Move oracle-only tests out of LibGit2 suite
Run version and version-height behavior tests against both managed and LibGit2 engines while retaining backend-specific lookup coverage.
